refactor(telegram): review fixes — uniform file IDs across the seam
Standards review: upload-tracking operations now take long file IDs like
every other seam operation, confining the TDLib int cast inside the real
adapter; the duplicated upload-release wait collapses into one
awaitUploadRelease helper used by both download paths. Spec review:
guessTypeFromMime documented as the no-message-context fallback (primary
classification lives once in the Telegram module) and a URL-construction
assertion pins the response contract. Suite: 44/44.

+++ src/main/java/com/zademy/nekolu/service/TelegramService.java
@@ -149,7 +149,7 @@ public interface TelegramService {
      * @param fileId the file ID
      * @param tempFilePath the staged file path
      */
-    void trackUpload(int fileId, String tempFilePath);
+    void trackUpload(long fileId, String tempFilePath);
 
     /**
      * Returns whether the file is still using a local staged source for upload.
@@ -157,7 +157,7 @@ public interface TelegramService {
      * @param fileId the file ID
      * @return true when the upload source is still tracked locally
      */
-    boolean isUploadTracked(int fileId);
+    boolean isUploadTracked(long fileId);
 
     /**
      * Waits until the staged local upload source for a file is released.
@@ -165,7 +165,7 @@ public interface TelegramService {
      * @param fileId the file ID
      * @return a future completed when the local upload source is no longer tracked
      */
-    CompletableFuture<Void> waitForUploadRelease(int fileId);
+    CompletableFuture<Void> waitForUploadRelease(long fileId);
 
     // ==================== FOLDERS ====================
